Optimize material shaders binding by caching time param once per draw pass
