+++ Source/Engine/Graphics/Materials/DeferredMaterialShader.cpp
@@ -14,7 +14,6 @@
 #include "Engine/Graphics/GPUDevice.h"
 #include "Engine/Graphics/Shaders/GPUShader.h"
 #include "Engine/Graphics/GPULimits.h"
-#include "Engine/Engine/Time.h"
 #include "Engine/Graphics/RenderTask.h"
 
 PACK_STRUCT(struct DeferredMaterialShaderData {
@@ -89,7 +88,7 @@ void DeferredMaterialShader::Bind(BindParameters& params)
         materialData->ViewPos = view.Position;
         materialData->ViewFar = view.Far;
         materialData->ViewDir = view.Direction;
-        materialData->TimeParam = Time::Draw.UnscaledTime.GetTotalSeconds();
+        materialData->TimeParam = params.TimeParam;
         materialData->ViewInfo = view.ViewInfo;
         materialData->ScreenSize = view.ScreenSize;
         const float scaleX = Vector3(drawCall.World.M11, drawCall.World.M12, drawCall.World.M13).Length();
