+++ Source/Engine/Graphics/Materials/DeformableMaterialShader.cpp
@@ -12,7 +12,6 @@
 #include "Engine/Graphics/GPUDevice.h"
 #include "Engine/Graphics/Shaders/GPUShader.h"
 #include "Engine/Graphics/GPULimits.h"
-#include "Engine/Engine/Time.h"
 #include "Engine/Graphics/RenderTask.h"
 
 PACK_STRUCT(struct DeformableMaterialShaderData {
@@ -76,7 +75,7 @@ void DeformableMaterialShader::Bind(BindParameters& params)
         materialData->ViewPos = view.Position;
         materialData->ViewFar = view.Far;
         materialData->ViewDir = view.Direction;
-        materialData->TimeParam = Time::Draw.UnscaledTime.GetTotalSeconds();
+        materialData->TimeParam = params.TimeParam;
         materialData->ViewInfo = view.ViewInfo;
         materialData->ScreenSize = view.ScreenSize;
         materialData->WorldDeterminantSign = drawCall.WorldDeterminantSign;
